pyproject.toml
[tool.poetry]
name = "hourly_wiki"
version = "0.1.0"
description = "twitter bot"
authors = ["Marie <schindler.marie@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
tweepy = "^4.10.0"
requests = "^2.28.0"
python-dotenv = "^0.20.0"
loguru = "^0.6.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
